Corrected GaxProperties and GaxGrpcProperties
It turns out, dependencies.properties isn't used only for dependency management, but also in runtime to determine gax, gax-grpc and even grpc versions that are sent in the request headers. This commit brings this file from the original gax-java project into this repo, and includes it in the build rules taht require it.

	modified:   BUILD

Added a reference to the new file so it is visible to other build rules.

	new file:   dependencies.properties

A partial copy of the file from gax-java.

	modified:   gax-grpc/BUILD
	modified:   gax/BUILD

Fixed build rules so the properties file is included in the artifact and accessible at runtime.

	modified:   gax/src/test/java/com/google/api/gax/core/GaxPropertiesTest.java
Fixed the test as now the property is available at runtime.