Replaced common protos with internal reference
	modified:   WORKSPACE

Updated WORKSPACE to import googleapis/googleapis as a repository, and referenced proto_library targets from that repository to build java_proto_library targets that are relevant to this repo. Had to fork googleapis/googleapis into my own github and remove language-specific rules from it. Otherwise I would have to import gapic-generator as well, which I was trying to avoid.

	new file:   common-protos/BUILD

Created build rules for the main modules imported into proto-google-common-protos. This is not fully representative of all the code copied by owlbot (see https://github.com/googleapis/java-common-protos/blob/main/.github/.OwlBot.yaml), this only represents the protos required to compile and test the targets in this repository for now.

	modified:   core/BUILD
	modified:   gax-grpc/BUILD
	modified:   gax-httpjson/BUILD
Replaced the maven dependency on external common protos artifact from the one managed within this repo. Verified that exporting a maven artifact for any of these targets generates a dependency block to the new common protos, in the same version as the the artifact being published.